# openapi.json
{
  "components": {
    "schemas": {
      "Aligned": {
        "properties": {
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null where this version does not carry the verse",
            "title": "Text"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "text"
        ],
        "title": "Aligned",
        "type": "object"
      },
      "AlignedVerse": {
        "properties": {
          "book": {
            "title": "Book",
            "type": "string"
          },
          "chapter": {
            "title": "Chapter",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "texts": {
            "items": {
              "$ref": "#/components/schemas/Aligned"
            },
            "title": "Texts",
            "type": "array"
          },
          "verse": {
            "title": "Verse",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "book",
          "chapter",
          "verse",
          "reference",
          "texts"
        ],
        "title": "AlignedVerse",
        "type": "object"
      },
      "BookOut": {
        "properties": {
          "abbreviation": {
            "title": "Abbreviation",
            "type": "string"
          },
          "chapters": {
            "title": "Chapters",
            "type": "integer"
          },
          "code": {
            "examples": [
              "PSA"
            ],
            "title": "Code",
            "type": "string"
          },
          "deuterocanonical": {
            "title": "Deuterocanonical",
            "type": "boolean"
          },
          "group": {
            "examples": [
              "WISDOM"
            ],
            "title": "Group",
            "type": "string"
          },
          "name": {
            "description": "In the language of the version",
            "title": "Name",
            "type": "string"
          },
          "testament": {
            "examples": [
              "OLD"
            ],
            "title": "Testament",
            "type": "string"
          }
        },
        "required": [
          "code",
          "name",
          "abbreviation",
          "testament",
          "group",
          "deuterocanonical",
          "chapters"
        ],
        "title": "BookOut",
        "type": "object"
      },
      "BookWhole": {
        "properties": {
          "book": {
            "$ref": "#/components/schemas/BookOut"
          },
          "chapters": {
            "items": {
              "$ref": "#/components/schemas/ChapterOfBook"
            },
            "title": "Chapters",
            "type": "array"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "book",
          "chapters"
        ],
        "title": "BookWhole",
        "type": "object"
      },
      "ChapterOfBook": {
        "properties": {
          "chapter": {
            "title": "Chapter",
            "type": "integer"
          },
          "numbering": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Numbering"
              },
              {
                "type": "null"
              }
            ]
          },
          "verses": {
            "items": {
              "$ref": "#/components/schemas/VerseOut"
            },
            "title": "Verses",
            "type": "array"
          }
        },
        "required": [
          "chapter",
          "numbering",
          "verses"
        ],
        "title": "ChapterOfBook",
        "type": "object"
      },
      "ChapterOut": {
        "properties": {
          "book": {
            "$ref": "#/components/schemas/BookOut"
          },
          "chapter": {
            "title": "Chapter",
            "type": "integer"
          },
          "next": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Neighbour"
              },
              {
                "type": "null"
              }
            ]
          },
          "numbering": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Numbering"
              },
              {
                "type": "null"
              }
            ]
          },
          "previous": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Neighbour"
              },
              {
                "type": "null"
              }
            ]
          },
          "verses": {
            "items": {
              "$ref": "#/components/schemas/VerseOut"
            },
            "title": "Verses",
            "type": "array"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "book",
          "chapter",
          "numbering",
          "previous",
          "next",
          "verses"
        ],
        "title": "ChapterOut",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/Problem"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "InputScheme": {
        "description": "Which numbering the caller wrote the reference in.",
        "enum": [
          "spine",
          "vulgate",
          "org",
          "douay"
        ],
        "title": "InputScheme",
        "type": "string"
      },
      "Neighbour": {
        "properties": {
          "book": {
            "title": "Book",
            "type": "string"
          },
          "chapter": {
            "title": "Chapter",
            "type": "integer"
          }
        },
        "required": [
          "book",
          "chapter"
        ],
        "title": "Neighbour",
        "type": "object"
      },
      "Numbering": {
        "description": "What another scheme calls this chapter.\n\nA string rather than a number, because the answer is sometimes a range. The\nVulgate psalm 9 is 9 and 10 elsewhere, and 114 and 115 are both 116.",
        "properties": {
          "counterpart": {
            "examples": [
              "9-10"
            ],
            "title": "Counterpart",
            "type": "string"
          },
          "scheme": {
            "$ref": "#/components/schemas/Scheme"
          }
        },
        "required": [
          "scheme",
          "counterpart"
        ],
        "title": "Numbering",
        "type": "object"
      },
      "PassageOut": {
        "properties": {
          "book": {
            "title": "Book",
            "type": "string"
          },
          "reference": {
            "examples": [
              "Jo 3,16-17"
            ],
            "title": "Reference",
            "type": "string"
          },
          "verses": {
            "items": {
              "$ref": "#/components/schemas/AlignedVerse"
            },
            "title": "Verses",
            "type": "array"
          },
          "versions": {
            "items": {
              "type": "string"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "reference",
          "book",
          "versions",
          "verses"
        ],
        "title": "PassageOut",
        "type": "object"
      },
      "Problem": {
        "properties": {
          "input": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "reason": {
            "$ref": "#/components/schemas/Reason"
          }
        },
        "required": [
          "reason",
          "message"
        ],
        "title": "Problem",
        "type": "object"
      },
      "Reason": {
        "description": "Why a request could not be answered. A closed set, and the contract.\n\nA caller branches on this. The message beside it is a courtesy and it is\nfree to change, which is why nothing should read it.\n\nThe last five are B1's orphan reasons, carried through rather than\ntranslated, so an address that fails to map says the same thing here as it\ndoes in `orphans.json`. A test asserts the two sets stay in step.",
        "enum": [
          "unknown_version",
          "unknown_book",
          "not_on_spine",
          "unpublished_in_version",
          "malformed",
          "whole_chapter",
          "range_too_large",
          "store_unavailable",
          "no_counterpart",
          "chapter_out_of_range",
          "verse_out_of_range",
          "psalm_title"
        ],
        "title": "Reason",
        "type": "string"
      },
      "ResolvedOut": {
        "properties": {
          "book": {
            "title": "Book",
            "type": "string"
          },
          "ids": {
            "examples": [
              [
                "1CO.13.4",
                "1CO.13.5"
              ]
            ],
            "items": {
              "type": "string"
            },
            "title": "Ids",
            "type": "array"
          },
          "preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The first verse of the span the default version publishes. Where that version has a gap it is a later verse, and the ids say which addresses the span covers.",
            "title": "Preview"
          },
          "reference": {
            "examples": [
              "1Cor 13,4-7"
            ],
            "title": "Reference",
            "type": "string"
          }
        },
        "required": [
          "reference",
          "book",
          "ids",
          "preview"
        ],
        "title": "ResolvedOut",
        "type": "object"
      },
...
      "Scheme": {
        "enum": [
          "vulgate",
          "org",
          "douay"
        ],
        "title": "Scheme",
        "type": "string"
      },
      "VerseOut": {
        "properties": {
          "book": {
            "title": "Book",
            "type": "string"
          },
          "chapter": {
            "title": "Chapter",
            "type": "integer"
          },
          "id": {
            "description": "The published verse id",
            "examples": [
              "PSA.50.3"
            ],
            "title": "Id",
            "type": "string"
          },
          "reference": {
            "examples": [
              "Sl 50,3"
            ],
            "title": "Reference",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "verse": {
            "title": "Verse",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "book",
          "chapter",
          "verse",
          "reference",
          "text"
        ],
        "title": "VerseOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "A read-only API for the Catholic Bible. 73 books, Portuguese, English and Latin, public domain throughout.",
    "title": "Catholic Bible",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/passage": {
      "get": {
        "operationId": "passage_v1_passage_get",
        "parameters": [
          {
            "in": "query",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "versions",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Versions"
            }
          },
          {
            "in": "query",
            "name": "scheme",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/InputScheme",
              "default": "spine"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such resource"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The reference cannot be resolved"
          }
        },
        "summary": "A written reference, read in one or more versions"
      }
    },
    "/v1/resolve": {
      "get": {
        "operationId": "resolve_v1_resolve_get",
        "parameters": [
          {
            "in": "query",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scheme",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/InputScheme",
              "default": "spine"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolvedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The reference cannot be resolved"
          }
        },
        "summary": "A written reference, resolved to addresses"
      }
    },
...
    "/v1/versions/{version}/books": {
      "get": {
        "operationId": "list_books_v1_versions__version__books_get",
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BookOut"
                  },
                  "title": "Response List Books V1 Versions  Version  Books Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such resource"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The reference cannot be resolved"
          }
        },
        "summary": "The books one translation reaches"
      }
    },
    "/v1/versions/{version}/books/{book}": {
      "get": {
        "operationId": "read_book_v1_versions__version__books__book__get",
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "book",
            "required": true,
            "schema": {
              "title": "Book",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookWhole"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such resource"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The reference cannot be resolved"
          }
        },
        "summary": "A whole book, grouped by chapter"
      }
    },
    "/v1/versions/{version}/books/{book}/chapters/{chapter}": {
      "get": {
        "operationId": "read_chapter_v1_versions__version__books__book__chapters__chapter__get",
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "book",
            "required": true,
            "schema": {
              "title": "Book",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "chapter",
            "required": true,
            "schema": {
              "title": "Chapter",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such resource"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The reference cannot be resolved"
          }
        },
        "summary": "One chapter, with its neighbours"
      }
    },
    "/v1/versions/{version}/books/{book}/chapters/{chapter}/verses/{verse}": {
      "get": {
        "operationId": "read_verse_v1_versions__version__books__book__chapters__chapter__verses__verse__get",
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "book",
            "required": true,
            "schema": {
              "title": "Book",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "chapter",
            "required": true,
            "schema": {
              "title": "Chapter",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "verse",
            "required": true,
            "schema": {
              "title": "Verse",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such resource"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The reference cannot be resolved"
          }
        },
        "summary": "One verse"
      }
    }
  }
}